+++ isce2_topsapp/__main__.py
@@ -332,7 +332,16 @@ def main():
         default="gunw_slc",
         help="Select the HyP3 entrypoint to use",
     )
+    parser.add_argument(
+        "++omp-num-threads",
+        type=int,
+        help="The number of OpenMP threads to use for parallel processing"
+    )
+
     args, unknowns = parser.parse_known_args()
+
+    if args.omp_num_threads:
+        os.environ["OMP_NUM_THREADS"] = str(args.omp_num_threads)
 
     sys.argv = [args.process, *unknowns]
     # FIXME: this gets better in python 3.10